Checkpoint

+++ Benchmarks/Suites.cs
@@ -6,15 +6,15 @@ using System.Threading.Tasks;
 using SimdDictionary;
 
 namespace Benchmarks {
-    public class BCLInsertion : Insertion<Dictionary<int, int>> {
+    public class BCLInsertion : Insertion<Dictionary<long, long>> {
     }
 
-    public class BCLLookup : Lookup<Dictionary<int, int>> { 
+    public class BCLLookup : Lookup<Dictionary<long, long>> { 
     }
 
-    public class SimdInsertion : Insertion<SimdDictionary<int, int>> {
+    public class SimdInsertion : Insertion<SimdDictionary<long, long>> {
     }
 
-    public class SimdLookup : Lookup<SimdDictionary<int, int>> { 
+    public class SimdLookup : Lookup<SimdDictionary<long, long>> { 
     }
 }
